Fix two CPU bugs that corrupted generation while every test passed
Checking output quality per quant -- rather than only tokens/sec -- found that
celeg's CPU backend emitted incoherent text for all 8 cached LFM2.5 files,
which the sweep reported as `ok` because they load and benchmark normally.
Both root causes are CPU-only; CUDA was correct throughout, verified against
an independent fp64 forward pass built from the GGUF itself.

- Attention query scale was applied twice on the query-key-norm path.
  cpu_gqa_decode_paged_parallel folds 1/sqrt(head_dim) into the scores, so
  apply_cpu_attention_qk must contribute only the ratio between the model's
  query scale and that default. The two non-QK-norm branches did; the two
  QK-norm branches multiplied by layout.query_scale itself, leaving every
  score 1/sqrt(head_dim) too small -- 8x for head_dim 64. Softmax over a
  single position normalises that away, so it was invisible at position 0 and
  only corrupted output once some position carried a much larger value vector
  than its neighbours, which is exactly what LFM2.5's massive activations do
  from layer 4 on. Only models with QK-norm were affected, which is why
  Nanbeige and MiniCPM5 looked fine. All four branches now derive the
  premultiplier from one expression.

- The AVX2 Q8_0 dot accumulated the scaled sub-block dot in an int.
  Each Q8_0 sub-block has its own fp16 scale of order 1e-3, so
  `static_cast<int>(d * dot)` discarded most of every sub-block. The scalar
  kernel already used a float. Q6_K nearby can accumulate in int only because
  its per-sub-block scale is an integer and its single float scale is applied
  once at the end.

Both were invisible to the existing tests, so both get a regression test that
was checked to fail when the fix is reverted:

- The Q8_0 dot cases all used d = 0x3c00 (fp16 1.0), which makes d*dot an
  exact integer and truncation a no-op. The new case uses realistic
  magnitudes -- scales near 1e-3 and small weights, so the scaled dot lands
  around 1 -- and shapes the weights like the activation so the dot cannot
  cancel to zero and hide under the tolerance floor. It covers Q4_0 and Q5_0
  on the same principle.
- attention_semantics_test now asserts the query premultiplier is identical
  with and without QK-norm for the same query_scale.

Also adds --dump-logits/--print-top to celeg-cpu-run, mirroring celeg-run, so
the two backends can be compared numerically on one checkpoint; the shared
implementation moves to celeg/app/logit_diagnostics.hpp instead of being
copied into the second app. Backend disagreement is the only reference
available for architectures no other engine can load.

Corrects docs/QUANTIZATION_SUPPORT_MATRIX.md, which recorded ncu as not
installed and deferred the GPU decode investigation on that basis. It is
installed at /usr/local/cuda-13.2/bin/ncu, just not on PATH. That section now
also records the sharper finding: celeg's decode rate does not respond to
resident weight bytes at all, which rules out a bandwidth ceiling.

CPU ctest 81/81 and CUDA ctest 91/91 pass. All 8 LFM2.5 files now generate
coherent text on CPU, matching CUDA, and Nanbeige-3B-Q8_0 improves too.

+++ src/backend/cpu/kernels/gguf_avx2.cpp
@@ -520,7 +520,13 @@ float cpu_gguf_dot_avx2(const std::byte* packed_row, GgmlType type,
         const auto* weights = reinterpret_cast<const BlockQ8_0*>(packed_row);
         for (size_t b = 0; b < blocks; ++b) {
             const CpuQ8KBlock& x = activation[b];
-            int block_total = 0;
+            // Each Q8_0 sub-block carries its own fp16 scale, so the running
+            // total is a float sum of scaled integer dots. It must not be an
+            // int: d is O(1e-3), so truncating d*dot per sub-block discards
+            // nearly the whole product. Q6_K above can accumulate in int only
+            // because its per-sub-block scale is an integer and the single
+            // float d is applied once at the end.
+            float block_total = 0.0f;
             for (int sub = 0; sub < 8; ++sub) {
                 const BlockQ8_0& weight = weights[b * 8 + static_cast<size_t>(sub)];
                 const float d = fp16_bits_to_float(weight.d);
@@ -536,10 +542,10 @@ float cpu_gguf_dot_avx2(const std::byte* packed_row, GgmlType type,
                     _mm256_extracti128_si256(a, 1));
                 const __m256i pl = _mm256_madd_epi16(wl, al);
                 const __m256i ph = _mm256_madd_epi16(wh, ah);
-                block_total += static_cast<int>(d * static_cast<float>(
-                    horizontal_sum_256(_mm256_add_epi32(pl, ph))));
+                block_total += d * static_cast<float>(
+                    horizontal_sum_256(_mm256_add_epi32(pl, ph)));
             }
-            total += x.d * static_cast<float>(block_total);
+            total += x.d * block_total;
         }
         return total;
     }
